Move _bundle_cls method for AnVIL tests; dissolve test case
